requirements.txt
Flask
flask-cors
scikit-learn
numpy
urllib3
gunicorn
requests
python-dotenv